moves PrivateKey to api
