+++ libraries/opensk/src/api/mod.rs
@@ -24,5 +24,6 @@ pub mod crypto;
 pub mod customization;
 pub mod firmware_protection;
 pub mod key_store;
+pub mod private_key;
 pub mod rng;
 pub mod user_presence;
