+++ libraries/opensk/src/ctap/credential_management.rs
@@ -353,12 +353,12 @@ pub fn process_credential_management<E: Env>(
 
 #[cfg(test)]
 mod test {
-    use super::super::crypto_wrapper::PrivateKey;
     use super::super::data_formats::{PinUvAuthProtocol, PublicKeyCredentialType};
     use super::super::pin_protocol::authenticate_pin_uv_auth_token;
     use super::super::CtapState;
     use super::*;
     use crate::api::crypto::ecdh::SecretKey as _;
+    use crate::api::private_key::PrivateKey;
     use crate::api::rng::Rng;
     use crate::env::test::TestEnv;
     use crate::env::EcdhSk;
